Created Resource bundle Refactored inline javascript and handlebar code out into separate files Precompiled handlebars code Upgraded handlebars to v3 Added files to plugin excludes General code cleanup
