+++ RestApiDocGrailsPlugin.groovy
@@ -1,19 +1,22 @@
 class RestApiDocGrailsPlugin {
     // the plugin version
-    def version = "0.6.1"
+    def version = "0.6.2"
     // the version or versions of Grails the plugin is designed for
     def grailsVersion = "2.3 > *"
     // resources that are excluded from plugin packaging
     def pluginExcludes = [
         "grails-app/views/error.gsp",
+        "grails-app/views/index.gsp",
         "grails-app/domain/org/restapidoc/test/TestObject.groovy",
         "grails-app/controllers/org/restapidoc/test/TestController.groovy",
-        "grails-app/controllers/org/restapidoc/test/AnotherTestController.groovy"
+        "grails-app/controllers/org/restapidoc/test/AnotherTestController.groovy",
+        "web-app/css/errors.css",
+        "web-app/css/main.css",
+        "web-app/css/mobile.css",
+        "web-app/images/**",
+        "web-app/js/application.js",
+        "web-app/js/jsondoc/templates/**"
     ]
-
-
-
-
 
     // TODO Fill in these fields
     def title = "RestApiDoc Plugin" // Headline display name of the plugin
@@ -37,7 +40,8 @@ This plugin allows you to document methods (description, HTTP path/verb, paramet
 //    def organization = [ name: "My Company", url: "http://www.my-company.com/" ]
 
     // Any additional developers beyond the author specified above.
-    def developers = [ [ name: "Benjamin Stevens", email: "b.stevens@ulg.ac.be" ]]
+    def developers = [ [ name: "Benjamin Stevens", email: "b.stevens@ulg.ac.be" ],
+    [ name: "Lawrence Lee", email: "valheru.ashen.shugar@gmail.com" ]]
 
     // Location of the plugin's issue tracker.
     def issueManagement = [ system: "GITHUB", url: "https://github.com/loic911/restapidoc/issues" ]
