+++ application.properties
@@ -1,5 +1,5 @@
 #Grails Metadata file
-#Fri Aug 08 14:43:13 CEST 2014
-app.grails.version=2.4.4
+#Mon Feb 09 10:17:59 CET 2015
+app.grails.version=2.3.11
 app.name=RestApiDoc
 app.servlet.version=2.5
