+++ grails-app/conf/BuildConfig.groovy
@@ -44,9 +44,8 @@ grails.project.dependency.resolution = {
 
     plugins {
         build ":tomcat:7.0.42"
-
         compile ":plugin-config:0.2.0"
-
+        runtime(':resources:1.2.8') { export = false }
 
         build(":release:3.0.1",
               ":rest-client-builder:1.0.3") {
