+++ plugin.xml
@@ -1,4 +1,4 @@
-<plugin name='rest-api-doc' version='0.6.1' grailsVersion='2.3 &gt; *'>
+<plugin name='rest-api-doc' version='0.6.2' grailsVersion='2.3 &gt; *'>
   <author>Loïc Rollus</author>
   <authorEmail>loicrollus@gmail.com</authorEmail>
   <title>RestApiDoc Plugin</title>
@@ -11,6 +11,7 @@ This plugin allows you to document methods (description, HTTP path/verb, paramet
   <resources>
     <resource>RestApiDocBootStrap</resource>
     <resource>RestApiDocDefaultConfig</resource>
+    <resource>RestApiDocResources</resource>
     <resource>org.restapidoc.RestApiDocController</resource>
   </resources>
 </plugin>
