+++ web-app/js/jsondoc/restapidoc-handlebars.js
@@ -1,0 +1,1 @@
+!function(){var n=Handlebars.template,a=Handlebars.templates=Handlebars.templates||{};a.apis=n({1:function(n,a,e,t){var l,s=a.helperMissing,r="function",i=this.escapeExpression;return'		<li><a href="#" id="'+i((l=null!=(l=a.jsondocId||(null!=n?n.jsondocId:n))?l:s,typeof l===r?l.call(n,{name:"jsondocId",hash:{},data:t}):l))+'" rel="api">'+i((l=null!=(l=a.name||(null!=n?n.name:n))?l:s,typeof l===r?l.call(n,{name:"name",hash:{},data:t}):l))+"</a></li>\n"},compiler:[6,">= 2.0.0-beta.1"],main:function(n,a,e,t){var l,s,r,i='<ul class="nav nav-list">\n	<li class="nav-header">APIs</li>\n';return s=null!=(s=a.apis||(null!=n?n.apis:n))?s:a.helperMissing,r={name:"apis",hash:{},fn:this.program(1,t,0),inverse:this.noop,data:t},l="function"==typeof s?s.call(n,r):s,a.apis||(l=a.blockHelperMissing.call(n,l,r)),null!=l&&(i+=l),i+"</ul>"},useData:!0}),a.main=n({compiler:[6,">= 2.0.0-beta.1"],main:function(n,a,e,t){var l,s=a.helperMissing,r="function",i=this.escapeExpression;return'<blockquote>\n  <p style="text-transform: uppercase;">API info</span></p>\n  <small>Base path: '+i((l=null!=(l=a.basePath||(null!=n?n.basePath:n))?l:s,typeof l===r?l.call(n,{name:"basePath",hash:{},data:t}):l))+"</small>\n  <small>Version: "+i((l=null!=(l=a.version||(null!=n?n.version:n))?l:s,typeof l===r?l.call(n,{name:"version",hash:{},data:t}):l))+"</small>\n</blockquote>"},useData:!0}),a.methods=n({1:function(n,a,e,t){var l,s,r=a.helperMissing,i="function",o=this.escapeExpression;return'    <div class="accordion-group">\n        <div class="accordion-heading">\n            <span class="label pull-right '+o((s=null!=(s=a.verb||(null!=n?n.verb:n))?s:r,typeof s===i?s.call(n,{name:"verb",hash:{},data:t}):s))+'" style="margin-right: 5px; margin-top: 8px;">'+o((s=null!=(s=a.verb||(null!=n?n.verb:n))?s:r,typeof s===i?s.call(n,{name:"verb",hash:{},data:t}):s))+'</span>\n            <a href="#_'+o((s=null!=(s=a.jsondocId||(null!=n?n.jsondocId:n))?s:r,typeof s===i?s.call(n,{name:"jsondocId",hash:{},data:t}):s))+'" id="'+o((s=null!=(s=a.jsondocId||(null!=n?n.jsondocId:n))?s:r,typeof s===i?s.call(n,{name:"jsondocId",hash:{},data:t}):s))+'" rel="method" data-parent="#accordion" data-toggle="collapse" class="accordion-toggle">'+o((s=null!=(s=a.path||(null!=n?n.path:n))?s:r,typeof s===i?s.call(n,{name:"path",hash:{},data:t}):s))+'</a>\n        </div>\n        <div class="accordion-body collapse" id="_'+o((s=null!=(s=a.jsondocId||(null!=n?n.jsondocId:n))?s:r,typeof s===i?s.call(n,{name:"jsondocId",hash:{},data:t}):s))+'">\n            <div class="accordion-inner">\n                <table class="table table-condensed table-striped table-bordered">\n                    <tr>\n                        <th style="width:15%;">Path</th>\n                        <td><code>'+o((s=null!=(s=a.path||(null!=n?n.path:n))?s:r,typeof s===i?s.call(n,{name:"path",hash:{},data:t}):s))+"</code></td>\n                    </tr>\n                    <tr>\n                        <th>Description</th>\n                        <td>"+o((s=null!=(s=a.description||(null!=n?n.description:n))?s:r,typeof s===i?s.call(n,{name:"description",hash:{},data:t}):s))+'</td>\n                    </tr>\n                    <tr>\n                        <th>Method</th>\n                        <td><span class="label '+o((s=null!=(s=a.verb||(null!=n?n.verb:n))?s:r,typeof s===i?s.call(n,{name:"verb",hash:{},data:t}):s))+'">'+o((s=null!=(s=a.verb||(null!=n?n.verb:n))?s:r,typeof s===i?s.call(n,{name:"verb",hash:{},data:t}):s))+"</span></td>\n                    </tr>\n"+(null!=(l=a["if"].call(n,null!=n?n.produces:n,{name:"if",hash:{},fn:this.program(2,t,0),inverse:this.noop,data:t}))?l:"")+(null!=(l=a["if"].call(n,null!=n?n.consumes:n,{name:"if",hash:{},fn:this.program(5,t,0),inverse:this.noop,data:t}))?l:"")+(null!=(l=a["if"].call(n,null!=n?n.headers:n,{name:"if",hash:{},fn:this.program(7,t,0),inverse:this.noop,data:t}))?l:"")+(null!=(l=a["if"].call(n,null!=n?n.pathparameters:n,{name:"if",hash:{},fn:this.program(10,t,0),inverse:this.noop,data:t}))?l:"")+(null!=(l=a["if"].call(n,null!=n?n.queryparameters:n,{name:"if",hash:{},fn:this.program(20,t,0),inverse:this.noop,data:t}))?l:"")+(null!=(l=a["if"].call(n,null!=n?n.bodyobject:n,{name:"if",hash:{},fn:this.program(22,t,0),inverse:this.noop,data:t}))?l:"")+(null!=(l=a["if"].call(n,null!=n?n.response:n,{name:"if",hash:{},fn:this.program(25,t,0),inverse:this.noop,data:t}))?l:"")+(null!=(l=a["if"].call(n,null!=n?n.apierrors:n,{name:"if",hash:{},fn:this.program(28,t,0),inverse:this.noop,data:t}))?l:"")+"                </table>\n            </div>\n        </div>\n    </div>\n"},2:function(n,a,e,t){var l;return"                    <tr>\n                        <th colspan=2>Produces</th>\n                    </tr>\n"+(null!=(l=a.each.call(n,null!=n?n.produces:n,{name:"each",hash:{},fn:this.program(3,t,0),inverse:this.noop,data:t}))?l:"")},3:function(n){return"                    <tr>\n                        <td colspan=2><code>"+this.escapeExpression(this.lambda(n,n))+"</code></td>\n                    </tr>\n"},5:function(n,a,e,t){var l;return"                    <tr>\n                        <th colspan=2>Consumes</th>\n                    </tr>\n"+(null!=(l=a.each.call(n,null!=n?n.consumes:n,{name:"each",hash:{},fn:this.program(3,t,0),inverse:this.noop,data:t}))?l:"")},7:function(n,a,e,t){var l;return"                    <tr>\n                        <th colspan=2>Headers</th>\n                    </tr>\n"+(null!=(l=a.each.call(n,null!=n?n.headers:n,{name:"each",hash:{},fn:this.program(8,t,0),inverse:this.noop,data:t}))?l:"")},8:function(n){var a=this.lambda,e=this.escapeExpression;return"                    <tr>\n                        <td><code>"+e(a(null!=n?n.name:n,n))+"</code></td>\n                        <td>"+e(a(null!=n?n.description:n,n))+"</td>\n                    </tr>\n"},10:function(n,a,e,t){var l;return"                    <tr>\n                        <th colspan=2>Path parameters</th>\n                    </tr>\n"+(null!=(l=a.each.call(n,null!=n?n.pathparameters:n,{name:"each",hash:{},fn:this.program(11,t,0),inverse:this.noop,data:t}))?l:"")},11:function(n,a,e,t){var l,s=this.lambda,r=this.escapeExpression;return"                    <tr>\n                        <td><code>"+r(s(null!=n?n.name:n,n))+"</code></td>\n                        <td>Required: "+r(s(null!=n?n.required:n,n))+"</td>\n\n                    </tr>\n                    <tr>\n                        <td></td>\n                        <td>Type: "+r(s(null!=n?n.type:n,n))+"</td>\n                    </tr>\n"+(null!=(l=a["if"].call(n,null!=n?n.description:n,{name:"if",hash:{},fn:this.program(12,t,0),inverse:this.noop,data:t}))?l:"")+(null!=(l=a["if"].call(n,null!=n?n.allowedvalues:n,{name:"if",hash:{},fn:this.program(14,t,0),inverse:this.noop,data:t}))?l:"")+(null!=(l=a["if"].call(n,null!=n?n.format:n,{name:"if",hash:{},fn:this.program(18,t,0),inverse:this.noop,data:t}))?l:"")},12:function(n){return"                    <tr>\n                        <td></td>\n                        <td>Description: "+this.escapeExpression(this.lambda(null!=n?n.description:n,n))+"</td>\n                    </tr>\n"},14:function(n,a,e,t){var l;return"                    <tr>\n                        <td></td>\n                        <td>Allowed values:\n"+(null!=(l=a.each.call(n,null!=n?n.allowedvalues:n,{name:"each",hash:{},fn:this.program(15,t,0),inverse:this.noop,data:t}))?l:"")+"                        </td>\n                    </tr>\n"},15:function(n,a,e,t){var l;return"                            "+this.escapeExpression(this.lambda(n,n))+(null!=(l=a.unless.call(n,t&&t.last,{name:"unless",hash:{},fn:this.program(16,t,0),inverse:this.noop,data:t}))?l:"")+"\n"},16:function(){return",&nbsp"},18:function(n){return"                    <tr>\n                        <td></td>\n                        <td>Format: "+this.escapeExpression(this.lambda(null!=n?n.format:n,n))+"</td>\n                    </tr>\n"},20:function(n,a,e,t){var l;return"                    <tr>\n                        <th colspan=2>Query parameters</th>\n                    </tr>\n"+(null!=(l=a.each.call(n,null!=n?n.queryparameters:n,{name:"each",hash:{},fn:this.program(11,t,0),inverse:this.noop,data:t}))?l:"")},22:function(n,a,e,t){var l;return"                    <tr>\n                        <th colspan=2>Body object</th>\n                    </tr>\n                    <tr>\n                        <td>Object</td>\n                        <td><code>"+this.escapeExpression(this.lambda(null!=(l=null!=n?n.bodyobject:n)?l.object:l,n))+"</code></td>\n                    </tr>\n"+(null!=(l=a["if"].call(n,null!=(l=null!=n?n.bodyobject:n)?l.map:l,{name:"if",hash:{},fn:this.program(23,t,0),inverse:this.noop,data:t}))?l:"")},23:function(n){var a,e=this.lambda,t=this.escapeExpression;return"                    <tr>\n                        <td>Map key</td>\n                        <td><code>"+t(e(null!=(a=null!=n?n.bodyobject:n)?a.mapKeyObject:a,n))+"</code></td>\n                    </tr>\n                    <tr>\n                        <td>Map value</td>\n                        <td><code>"+t(e(null!=(a=null!=n?n.bodyobject:n)?a.mapValueObject:a,n))+"</code></td>\n                    </tr>\n"},25:function(n,a,e,t){var l;return"                    <tr>\n                        <th colspan=2>Response object</th>\n                    </tr>\n                    <tr>\n                        <td>Object</td>\n                        <td><code>"+this.escapeExpression(this.lambda(null!=(l=null!=n?n.response:n)?l.object:l,n))+"</code></td>\n                    </tr>\n"+(null!=(l=a["if"].call(n,null!=(l=null!=n?n.response:n)?l.map:l,{name:"if",hash:{},fn:this.program(26,t,0),inverse:this.noop,data:t}))?l:"")},26:function(n){var a,e=this.lambda,t=this.escapeExpression;return"                    <tr>\n                        <td>Map key</td>\n                        <td><code>"+t(e(null!=(a=null!=n?n.response:n)?a.mapKeyObject:a,n))+"</code></td>\n                    </tr>\n                    <tr>\n                        <td>Map value</td>\n                        <td><code>"+t(e(null!=(a=null!=n?n.response:n)?a.mapValueObject:a,n))+"</code></td>\n                    </tr>\n"},28:function(n,a,e,t){var l;return"                    <tr>\n                        <th colspan=2>Errors</th>\n                    </tr>\n"+(null!=(l=a.each.call(n,null!=n?n.apierrors:n,{name:"each",hash:{},fn:this.program(29,t,0),inverse:this.noop,data:t}))?l:"")},29:function(n){var a=this.lambda,e=this.escapeExpression;return"                    <tr>\n                        <td><code>"+e(a(null!=n?n.code:n,n))+"</code></td>\n                        <td>"+e(a(null!=n?n.description:n,n))+"</td>\n                    </tr>\n"},compiler:[6,">= 2.0.0-beta.1"],main:function(n,a,e,t){var l,s,r,i='<blockquote>\n  <p style="text-transform: uppercase;"><span id="apiName"></span></p>\n  <small><span id="apiDescription"></span></cite></small>\n</blockquote>\n\n<div class="accordion" id="accordion">\n';return s=null!=(s=a.methods||(null!=n?n.methods:n))?s:a.helperMissing,r={name:"methods",hash:{},fn:this.program(1,t,0),inverse:this.noop,data:t},l="function"==typeof s?s.call(n,r):s,a.methods||(l=a.blockHelperMissing.call(n,l,r)),null!=l&&(i+=l),i+"</div>"},useData:!0}),a.object=n({1:function(n,a,e,t){var l;return"    <tr><th>Description</th><td>"+this.escapeExpression((l=null!=(l=a.description||(null!=n?n.description:n))?l:a.helperMissing,"function"==typeof l?l.call(n,{name:"description",hash:{},data:t}):l))+"</td></tr>\n"},3:function(n,a,e,t){var l;return"    <tr><th colspan=2>Fields</th></tr>\n"+(null!=(l=a.each.call(n,null!=n?n.fields:n,{name:"each",hash:{},fn:this.program(4,t,0),inverse:this.noop,data:t}))?l:"")},4:function(n,a,e,t){var l,s,r=a.helperMissing,i="function",o=this.escapeExpression;return"    <tr><td><code>"+o((s=null!=(s=a.name||(null!=n?n.name:n))?s:r,typeof s===i?s.call(n,{name:"name",hash:{},data:t}):s))+"</code></td><td>"+o((s=null!=(s=a.description||(null!=n?n.description:n))?s:r,typeof s===i?s.call(n,{name:"description",hash:{},data:t}):s))+"</td></tr>\n    <tr><td></td><td>Type: "+o((s=null!=(s=a.type||(null!=n?n.type:n))?s:r,typeof s===i?s.call(n,{name:"type",hash:{},data:t}):s))+"</td></tr>\n    <tr><td></td><td>Multiple: "+o((s=null!=(s=a.multiple||(null!=n?n.multiple:n))?s:r,typeof s===i?s.call(n,{name:"multiple",hash:{},data:t}):s))+"</td></tr>\n"+(null!=(l=a["if"].call(n,null!=n?n.useForCreation:n,{name:"if",hash:{},fn:this.program(5,t,0),inverse:this.noop,data:t}))?l:"")+"\n"},5:function(n,a,e,t){var l;return'    <tr>\n        <td></td>\n        <td><span class="label label-info">Use for creation</span></td>\n    </tr>\n'+(null!=(l=a["if"].call(n,null!=n?n.mandatory:n,{name:"if",hash:{},fn:this.program(6,t,0),inverse:this.program(8,t,0),data:t}))?l:"")+(null!=(l=a["if"].call(n,null!=n?n.presentInResponse:n,{name:"if",hash:{},fn:this.program(10,t,0),inverse:this.program(12,t,0),data:t}))?l:"")+"\n"},6:function(){return'    <tr>\n        <td></td>\n        <td><span class="label label-warning">Mandatory</span></td>\n    </tr>\n'},8:function(n,a,e,t){var l;return'    <tr>\n        <td></td>\n        <td><span class="label label-warning">Default: '+this.escapeExpression((l=null!=(l=a.defaultValue||(null!=n?n.defaultValue:n))?l:a.helperMissing,"function"==typeof l?l.call(n,{name:"defaultValue",hash:{},data:t}):l))+"</span></td>\n    </tr>\n"},10:function(){return""},12:function(){return'    <tr>\n        <td></td>\n        <td><span class="label label-inverse">Not in response!</span></td>\n    </tr>\n'},compiler:[6,">= 2.0.0-beta.1"],main:function(n,a,e,t){var l,s;return'<table class="table table-condensed table-striped table-bordered">\n    <tr><th style="width:15%;">Name</th><td><code>'+this.escapeExpression((s=null!=(s=a.name||(null!=n?n.name:n))?s:a.helperMissing,"function"==typeof s?s.call(n,{name:"name",hash:{},data:t}):s))+"</code></td></tr>\n"+(null!=(l=a["if"].call(n,null!=n?n.description:n,{name:"if",hash:{},fn:this.program(1,t,0),inverse:this.noop,data:t}))?l:"")+(null!=(l=a["if"].call(n,null!=n?n.fields:n,{name:"if",hash:{},fn:this.program(3,t,0),inverse:this.noop,data:t}))?l:"")+"</table>"},useData:!0}),a.objects=n({1:function(n,a,e,t){var l,s=a.helperMissing,r="function",i=this.escapeExpression;return'		<li><a href="#" id="'+i((l=null!=(l=a.jsondocId||(null!=n?n.jsondocId:n))?l:s,typeof l===r?l.call(n,{name:"jsondocId",hash:{},data:t}):l))+'" rel="object">'+i((l=null!=(l=a.name||(null!=n?n.name:n))?l:s,typeof l===r?l.call(n,{name:"name",hash:{},data:t}):l))+"</a></li>\n"},compiler:[6,">= 2.0.0-beta.1"],main:function(n,a,e,t){var l,s,r,i='<ul class="nav nav-list">\n	<li class="nav-header">Objects</li>\n';return s=null!=(s=a.objects||(null!=n?n.objects:n))?s:a.helperMissing,r={name:"objects",hash:{},fn:this.program(1,t,0),inverse:this.noop,data:t},l="function"==typeof s?s.call(n,r):s,a.objects||(l=a.blockHelperMissing.call(n,l,r)),null!=l&&(i+=l),i+"</ul>"},useData:!0}),a.test=n({1:function(n,a,e,t){var l,s,r,i='    <div class="span12">\n        <div id="headers">\n            <h4>Headers</h4>\n';return s=null!=(s=a.headers||(null!=n?n.headers:n))?s:a.helperMissing,r={name:"headers",hash:{},fn:this.program(2,t,0),inverse:this.noop,data:t},l="function"==typeof s?s.call(n,r):s,a.headers||(l=a.blockHelperMissing.call(n,l,r)),null!=l&&(i+=l),i+"        </div>\n    </div>\n"},2:function(n,a,e,t){var l,s=a.helperMissing,r="function",i=this.escapeExpression;return'            <div class="input-prepend">\n                <span style="text-align:left;" class="add-on span4">'+i((l=null!=(l=a.name||(null!=n?n.name:n))?l:s,typeof l===r?l.call(n,{name:"name",hash:{},data:t}):l))+'</span><input type="text" class="span8" name="'+i((l=null!=(l=a.name||(null!=n?n.name:n))?l:s,typeof l===r?l.call(n,{name:"name",hash:{},data:t}):l))+'" placeholder="'+i((l=null!=(l=a.name||(null!=n?n.name:n))?l:s,typeof l===r?l.call(n,{name:"name",hash:{},data:t}):l))+'">\n            </div>\n'},4:function(n,a,e,t){var l,s,r,i='    <div class="span6" style="margin-left:0px">\n        <div id="produces" class="playground-spacer">\n            <h4>Accept</h4>\n';return s=null!=(s=a.produces||(null!=n?n.produces:n))?s:a.helperMissing,r={name:"produces",hash:{},fn:this.program(5,t,0),inverse:this.noop,data:t},l="function"==typeof s?s.call(n,r):s,a.produces||(l=a.blockHelperMissing.call(n,l,r)),null!=l&&(i+=l),i+"        </div>\n    </div>\n"},5:function(n){var a=this.lambda,e=this.escapeExpression;return'            <label class="radio"><input type="radio" name="produces" value="'+e(a(n,n))+'">'+e(a(n,n))+"</label>\n"},7:function(n,a,e,t){var l;return null!=(l=a["if"].call(n,null!=n?n.consumes:n,{name:"if",hash:{},fn:this.program(8,t,0),inverse:this.noop,data:t}))?l:""},8:function(n,a,e,t){var l,s,r,i='    <div class="span6" style="margin-left:0px">\n        <div id="consumes" class="playground-spacer">\n            <h4>Content type</h4>\n';return s=null!=(s=a.consumes||(null!=n?n.consumes:n))?s:a.helperMissing,r={name:"consumes",hash:{},fn:this.program(9,t,0),inverse:this.noop,data:t},l="function"==typeof s?s.call(n,r):s,a.consumes||(l=a.blockHelperMissing.call(n,l,r)),null!=l&&(i+=l),i+"        </div>\n    </div>\n"},9:function(n){var a=this.lambda,e=this.escapeExpression;return'            <label class="radio"><input type="radio" name="consumes" value="'+e(a(n,n))+'">'+e(a(n,n))+"</label>\n"},11:function(n,a,e,t){var l,s,r,i='    <div class="span12" style="margin-left:0px">\n        <div id="pathparameters" class="playground-spacer">\n            <h4>Path parameters</h4>\n';return s=null!=(s=a.pathparameters||(null!=n?n.pathparameters:n))?s:a.helperMissing,r={name:"pathparameters",hash:{},fn:this.program(2,t,0),inverse:this.noop,data:t},l="function"==typeof s?s.call(n,r):s,a.pathparameters||(l=a.blockHelperMissing.call(n,l,r)),null!=l&&(i+=l),i+"        </div>\n    </div>\n"},13:function(n,a,e,t){var l,s,r,i='    <div class="span12" style="margin-left:0px">\n        <div id="queryparameters" class="playground-spacer">\n            <h4>Query parameters</h4>\n';return s=null!=(s=a.queryparameters||(null!=n?n.queryparameters:n))?s:a.helperMissing,r={name:"queryparameters",hash:{},fn:this.program(2,t,0),inverse:this.noop,data:t},l="function"==typeof s?s.call(n,r):s,a.queryparameters||(l=a.blockHelperMissing.call(n,l,r)),null!=l&&(i+=l),i+"        </div>\n    </div>\n"},15:function(){return'    <div class="span12" style="margin-left:0px">\n        <div id="bodyobject" class="playground-spacer">\n            <h4>Body object</h4>\n            <label class="select">Body type:\n                <g:select name="bodyTypeSelect" \n                from="${[\'String body\', \'Multipart body\']}"\n                value="String body"\n                onchange="updateMethodBodyForm();">\n            </g:select>\n            <label>\n                <div id="methodStringBody">\n                    <textarea class="span12" id="inputJson" rows=10 />\n                </div>\n                <div id="methodMultipartBody" style="display:none;">\n                    <g:form name="multipartFilesForm" enctype="multipart/form-data">\n                    Files: <input type="file" name="inputMultipartFiles" onchange="chooseFileUpdate(this.files);" multiple="true"/>\n                </g:form>\n            </div>\n        </div>\n    </div>\n'},compiler:[6,">= 2.0.0-beta.1"],main:function(n,a,e,t){var l,s;return'<blockquote>\n  <p style="text-transform: uppercase;">Playground</span></p>\n  <small>'+this.escapeExpression((s=null!=(s=a.path||(null!=n?n.path:n))?s:a.helperMissing,"function"==typeof s?s.call(n,{name:"path",hash:{},data:t}):s))+'</small>\n</blockquote>\n\n<div class="row-fluid">\n\n'+(null!=(l=a["if"].call(n,null!=n?n.headers:n,{name:"if",hash:{},fn:this.program(1,t,0),inverse:this.noop,data:t}))?l:"")+"\n"+(null!=(l=a["if"].call(n,null!=n?n.produces:n,{name:"if",hash:{},fn:this.program(4,t,0),inverse:this.noop,data:t}))?l:"")+"\n"+(null!=(l=a["if"].call(n,null!=n?n.bodyobject:n,{name:"if",hash:{},fn:this.program(7,t,0),inverse:this.noop,data:t}))?l:"")+"\n"+(null!=(l=a["if"].call(n,null!=n?n.pathparameters:n,{name:"if",hash:{},fn:this.program(11,t,0),inverse:this.noop,data:t}))?l:"")+"\n"+(null!=(l=a["if"].call(n,null!=n?n.queryparameters:n,{name:"if",hash:{},fn:this.program(13,t,0),inverse:this.noop,data:t}))?l:"")+"\n"+(null!=(l=a["if"].call(n,null!=n?n.bodyobject:n,{name:"if",hash:{},fn:this.program(15,t,0),inverse:this.noop,data:t}))?l:"")+'\n    <div class="span12" style="margin-left:0px">\n        <div class="form-actions">\n            <button class="btn btn-primary" id="testButton" data-loading-text="Loading...">Submit</button>\n        </div>\n    </div>\n\n</div>\n\n<div class="tabbable" id="resInfo" style="display:none;">\n    <ul class="nav nav-tabs">\n      <li class="active"><a href="#tab1" data-toggle="tab">Response text</a></li>\n      <li><a href="#tab2" data-toggle="tab">Response info</a></li>\n      <li><a href="#tab3" data-toggle="tab">Request info</a></li>\n  </ul>\n  <div class="tab-content">\n    <div class="tab-pane active" id="tab1">\n        <pre id="response" class="prettyprint">\n        </pre>\n    </div>\n    <div class="tab-pane" id="tab2">\n        <p class="nav-header" style="padding:0px">Response code</p>\n        <pre id="responseStatus" class="prettyprint">\n        </pre>\n        <p class="nav-header" style="padding:0px">Response headers</p>\n        <pre id="responseHeaders" class="prettyprint">\n        </pre>\n    </div>\n    <div class="tab-pane" id="tab3">\n      <p class="nav-header" style="padding:0px">Request URL</p>\n      <pre id="requestURL" class="prettyprint">\n      </pre>\n  </div>\n</div>\n</div>'},useData:!0})}();
