+++ web-app/js/jsondoc/templates/apis.handlebars
@@ -1,0 +1,6 @@
+<ul class="nav nav-list">
+	<li class="nav-header">APIs</li>
+	{{#apis}}
+		<li><a href="#" id="{{jsondocId}}" rel="api">{{name}}</a></li>
+	{{/apis}}
+</ul>
