+++ web-app/js/jsondoc/templates/main.handlebars
@@ -1,0 +1,5 @@
+<blockquote>
+  <p style="text-transform: uppercase;">API info</span></p>
+  <small>Base path: {{basePath}}</small>
+  <small>Version: {{version}}</small>
+</blockquote>
