+++ web-app/js/jsondoc/templates/objects.handlebars
@@ -1,0 +1,6 @@
+<ul class="nav nav-list">
+	<li class="nav-header">Objects</li>
+	{{#objects}}
+		<li><a href="#" id="{{jsondocId}}" rel="object">{{name}}</a></li>
+	{{/objects}}
+</ul>
