+++ web-app/js/jsondoc/templates/test.handlebars
@@ -1,0 +1,127 @@
+<blockquote>
+  <p style="text-transform: uppercase;">Playground</span></p>
+  <small>{{path}}</small>
+</blockquote>
+
+<div class="row-fluid">
+
+    {{#if headers}}
+    <div class="span12">
+        <div id="headers">
+            <h4>Headers</h4>
+            {{#headers}}
+            <div class="input-prepend">
+                <span style="text-align:left;" class="add-on span4">{{name}}</span><input type="text" class="span8" name="{{name}}" placeholder="{{name}}">
+            </div>
+            {{/headers}}
+        </div>
+    </div>
+    {{/if}}
+
+    {{#if produces}}
+    <div class="span6" style="margin-left:0px">
+        <div id="produces" class="playground-spacer">
+            <h4>Accept</h4>
+            {{#produces}}
+            <label class="radio"><input type="radio" name="produces" value="{{this}}">{{this}}</label>
+            {{/produces}}
+        </div>
+    </div>
+    {{/if}}
+
+    {{#if bodyobject}}
+    {{#if consumes}}
+    <div class="span6" style="margin-left:0px">
+        <div id="consumes" class="playground-spacer">
+            <h4>Content type</h4>
+            {{#consumes}}
+            <label class="radio"><input type="radio" name="consumes" value="{{this}}">{{this}}</label>
+            {{/consumes}}
+        </div>
+    </div>
+    {{/if}}
+    {{/if}}
+
+    {{#if pathparameters}}
+    <div class="span12" style="margin-left:0px">
+        <div id="pathparameters" class="playground-spacer">
+            <h4>Path parameters</h4>
+            {{#pathparameters}}
+            <div class="input-prepend">
+                <span style="text-align:left;" class="add-on span4">{{name}}</span><input type="text" class="span8" name="{{name}}" placeholder="{{name}}">
+            </div>
+            {{/pathparameters}}
+        </div>
+    </div>
+    {{/if}}
+
+    {{#if queryparameters}}
+    <div class="span12" style="margin-left:0px">
+        <div id="queryparameters" class="playground-spacer">
+            <h4>Query parameters</h4>
+            {{#queryparameters}}
+            <div class="input-prepend">
+                <span style="text-align:left;" class="add-on span4">{{name}}</span><input type="text" class="span8" name="{{name}}" placeholder="{{name}}">
+            </div>
+            {{/queryparameters}}
+        </div>
+    </div>
+    {{/if}}
+
+    {{#if bodyobject}}
+    <div class="span12" style="margin-left:0px">
+        <div id="bodyobject" class="playground-spacer">
+            <h4>Body object</h4>
+            <label class="select">Body type:
+                <g:select name="bodyTypeSelect" 
+                from="${['String body', 'Multipart body']}"
+                value="String body"
+                onchange="updateMethodBodyForm();">
+            </g:select>
+            <label>
+                <div id="methodStringBody">
+                    <textarea class="span12" id="inputJson" rows=10 />
+                </div>
+                <div id="methodMultipartBody" style="display:none;">
+                    <g:form name="multipartFilesForm" enctype="multipart/form-data">
+                    Files: <input type="file" name="inputMultipartFiles" onchange="chooseFileUpdate(this.files);" multiple="true"/>
+                </g:form>
+            </div>
+        </div>
+    </div>
+    {{/if}}
+
+    <div class="span12" style="margin-left:0px">
+        <div class="form-actions">
+            <button class="btn btn-primary" id="testButton" data-loading-text="Loading...">Submit</button>
+        </div>
+    </div>
+
+</div>
+
+<div class="tabbable" id="resInfo" style="display:none;">
+    <ul class="nav nav-tabs">
+      <li class="active"><a href="#tab1" data-toggle="tab">Response text</a></li>
+      <li><a href="#tab2" data-toggle="tab">Response info</a></li>
+      <li><a href="#tab3" data-toggle="tab">Request info</a></li>
+  </ul>
+  <div class="tab-content">
+    <div class="tab-pane active" id="tab1">
+        <pre id="response" class="prettyprint">
+        </pre>
+    </div>
+    <div class="tab-pane" id="tab2">
+        <p class="nav-header" style="padding:0px">Response code</p>
+        <pre id="responseStatus" class="prettyprint">
+        </pre>
+        <p class="nav-header" style="padding:0px">Response headers</p>
+        <pre id="responseHeaders" class="prettyprint">
+        </pre>
+    </div>
+    <div class="tab-pane" id="tab3">
+      <p class="nav-header" style="padding:0px">Request URL</p>
+      <pre id="requestURL" class="prettyprint">
+      </pre>
+  </div>
+</div>
+</div>
